+++ Data/SavedData.lua
@@ -52,7 +52,7 @@ end
 
 local FixFeaturesTutorialProgress, FixElvUISkin, FixFilters, FixEventDetails, FixShowExcludedCategory, FixEventDetails2, FixCharacters, FixEventAlert;
 local FixMergeSmallCategoriesThresholdChanged, FixShowCurrentCharacterIcons, FixTabs, FixCovenantFilters, FixNewEarnedByFilter, FixTabs2, FixNewEarnedByFilter2;
-local FixEventDetails3, FixTooltipCriteria, FixFocusedAchievements, FixFocusedOptions, FixEventRemindersTimeDisplay;
+local FixEventDetails3, FixTooltipCriteria, FixFocusedAchievements, FixFocusedOptions, FixEventRemindersTimeDisplay, FixEventRemindersShowPopUpsOptions;
 function LoadSolutions()
     local solutions = {
         FixFeaturesTutorialProgress, -- 1
@@ -75,6 +75,7 @@ function LoadSolutions()
         FixFocusedAchievements, -- 18
         FixFocusedOptions, -- 19
         FixEventRemindersTimeDisplay, -- 20
+        FixEventRemindersShowPopUpsOptions, -- 21
     };
 
     return solutions;
@@ -593,4 +594,24 @@ function FixEventRemindersTimeDisplay(prevBuild, currBuild, prevVersion, currVer
     SavedData.Fixes.FixEventRemindersTimeDisplay = true;
     
     diagnostics.Debug("EventReminders TimeDisplay Options fixed");
+end
+
+function FixEventRemindersShowPopUpsOptions(prevBuild, currBuild, prevVersion, currVersion, firstTime)
+    -- In version 51.0 addon.Options.db.EventReminders.ShowPopUps was moved to addon.Options.db.EventReminders.ShowPopUps.OnLogin
+    -- Here we clean up the old addon.Options.db.EventReminders.ShowPopUps for users pre 51.0
+    -- addon.Options.db.EventReminders.ShowPopUps.OnLogin is created by the Options so we don't need to do this here, just copy if previous existed
+
+    if firstTime and currVersion > "51.0" then
+        diagnostics.Debug("First time EventReminders ShowPopUps Options OK");
+        return;
+    end
+    if addon.Options.db.EventReminders.ShowPopUps == nil then
+        diagnostics.Debug("EventReminders ShowPopUps Options already moved");
+        return;
+    end
+
+    addon.Options.db.EventReminders.ShowPopUps.OnLogin = addon.Options.db.EventReminders.ShowPopUps;
+    addon.Options.db.EventReminders.ShowPopUps = nil;
+
+    diagnostics.Debug("EventReminders ShowPopUps Options moved");
 end